requirements.txt
torch==1.6.0
numpy>=1.14.5
cn2an==0.5.12
ruamel.yaml==0.16.5
torchmetrics==0.7.2
sanic==22.3.0
aiohttp==3.7.4
python-socketio==5.5.2
python-engineio==4.3.1
